Rename error() method to getError()
When both a method and a signal have the same name (e.g., QNetworkReply::error() const
is a method and QNetworkReply::error(QNetworkReply::NetworkError) is a signal) and the
signal is emitted, the script does not see that signal. In the example with QNetworkReply
the script will never be notified of an error.

The solution is to rename the method. In this case, error() method was renamed to getError().